+++ CHANGELOG.md
@@ -3,6 +3,31 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v2.8.0
+------------------------------
+*October 23, 2020*
+
+### Added
+- Stylelint added to lint styling on component builds (for relevant packages).
+
+
+v2.7.0
+------------------------------
+*October 21, 2020*
+
+### Added
+- Added `#globalconfig` flag that can be added to PR titles to stop danger running on all packages (only runs on root).
+
+
+v2.6.0
+------------------------------
+*October 21, 2020*
+
+### Added
+- Upgrading `fozzie-colour-palette` and `fozzie` package versions.
+
+
 v2.5.0
 ------------------------------
 *October 15, 2020*
@@ -10,6 +35,7 @@ v2.5.0
 ### Changed
 - Circle CI Docker image
 - Chromedriver version to v86.0.0
+
 
 v2.4.0
 ------------------------------